ApplicationControllerの書き換え
