+++ app/controllers/customer/top_controller.rb
@@ -1,6 +1,5 @@
 class Customer::TopController < ApplicationController
   def index
-    raise ActiveRecord::RecordNotFound
     render action: "index"
   end
 end
